# api-docs/swagger.yaml
basePath: /api/v1
definitions:
  expense.CategoryTotal:
    properties:
      category:
        example: food
        type: string
      total:
        example: 15075
        type: integer
    type: object
  expense.CreateExpenseRequest:
    properties:
      amount:
        type: integer
      category:
        type: string
      date:
        type: string
      title:
        type: string
    required:
    - amount
    - category
    - date
    - title
    type: object
  expense.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: Invalid input
        type: string
    type: object
  expense.Expense:
    properties:
      amount:
        example: 2550
        type: integer
      category:
        example: food
        type: string
      date:
        example: "2026-08-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      title:
        example: Lunch at restaurant
        type: string
    type: object
  expense.MonthlySummary:
    properties:
      by_category:
        additionalProperties:
          format: int64
          type: integer
        type: object
      month:
        example: 2026-08
        type: string
      total:
        example: 50000
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@expensesync.io
    name: Support
  description: A smart REST API for managing personal expenses
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: ExpenseSync API
  version: 1.0.0
paths:
  /expenses:
    get:
      description: Retrieve all expenses with optional filtering by category or title
        search
      parameters:
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Search by title
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/expense.Expense'
            type: array
      summary: Get all expenses
      tags:
      - expenses
    post:
      consumes:
      - application/json
      description: Add a new expense to the tracker
      parameters:
      - description: Expense details
        in: body
        name: expense
        required: true
        schema:
          $ref: '#/definitions/expense.CreateExpenseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/expense.Expense'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/expense.ErrorResponse'
      summary: Create new expense
      tags:
      - expenses
  /expenses/{id}:
    delete:
      description: Remove an expense by its unique ID
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/expense.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/expense.ErrorResponse'
      summary: Delete expense
      tags:
      - expenses
    get:
      description: Retrieve a specific expense by its unique ID
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/expense.Expense'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/expense.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/expense.ErrorResponse'
      summary: Get expense by ID
      tags:
      - expenses
  /expenses/category/{category}/total:
    get:
      description: Calculate total expenses for a specific category
      parameters:
      - description: Category name
        in: path
        name: category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/expense.CategoryTotal'
      summary: Get category total
      tags:
      - calculations
  /expenses/totals:
    get:
      description: Calculate overall total of all expenses
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              format: int64
              type: integer
            type: object
      summary: Get total expenses
      tags:
      - calculations
  /summary/monthly:
    get:
      description: Get total expenses grouped by month and category
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/expense.MonthlySummary'
            type: array
      summary: Get monthly summary
      tags:
      - analytics
swagger: "2.0"